Move module/ into src/
