+++ src/module/conflict/conflict.js
