+++ src/module/conflict/dispoDialog.js
