+++ src/module/conflict/types.js
