+++ src/module/inventory/containerValidator.js
