+++ src/module/inventory/inventory.js
