+++ src/module/rolls/helpingDiceDialog.js
