+++ src/module/rolls/playerRollDialog.js
